Predicting US Presidential Election Result Using Twitter Sentiment Analysis
According to the reports of “The New York Times”, mostly everyone has dropped out till April 2020 who was running for the Presidential election and the only left ones are Donald Trump and Joe Biden now. So, I have done a small analysis for trying to predict the winner of this election.